コンポーネント作成のため、自作ファイルをoldにして、教材のファイルをコピーしてきました
