+++ src/main.js
@@ -3,7 +3,7 @@ import './assets/main.css'
 import { createApp } from 'vue'
 import { createPinia } from 'pinia'
 
-import App from './App.vue'
+import App from './App(old).vue'
 import router from './router'
 
 const app = createApp(App)
